refactor(config): move static methods from HoconConfigModule to HoconConfigFactory
Move extractIncludedFiles, collectFileOrigins, enrichOriginWithIncludes
from the interface to the factory class where they belong.
