+++ config/config-hocon/src/main/java/ru/tinkoff/kora/config/hocon/HoconConfigModule.java
@@ -13,11 +13,7 @@ import ru.tinkoff.kora.config.common.origin.ResourceConfigOrigin;
 import ru.tinkoff.kora.config.common.origin.SimpleConfigOrigin;
 
 import java.net.URISyntaxException;
-import java.nio.file.Files;
 import java.nio.file.Path;
-import java.util.ArrayList;
-import java.util.LinkedHashSet;
-import java.util.Set;
 
 public interface HoconConfigModule extends CommonConfigModule {
     @ApplicationConfig
@@ -46,12 +42,12 @@ public interface HoconConfigModule extends CommonConfigModule {
             }
             if (resourceUrl.getProtocol().equals("file")) {
                 var path = Path.of(resourceUrl.toURI());
-                return enrichOriginWithIncludes(new FileConfigOrigin(path), ConfigFactory.parseFile(path.toFile()));
+                return HoconConfigFactory.enrichOriginWithIncludes(new FileConfigOrigin(path), ConfigFactory.parseFile(path.toFile()));
             }
             return new ResourceConfigOrigin(resourceUrl);
         } else {
             var path = Path.of(file);
-            return enrichOriginWithIncludes(new FileConfigOrigin(path), ConfigFactory.parseFile(path.toFile()));
+            return HoconConfigFactory.enrichOriginWithIncludes(new FileConfigOrigin(path), ConfigFactory.parseFile(path.toFile()));
         }
     }
 
@@ -85,52 +81,5 @@ public interface HoconConfigModule extends CommonConfigModule {
     @ApplicationConfig
     default Config config(@ApplicationConfig ConfigOrigin origin, com.typesafe.config.Config hoconConfig) {
         return HoconConfigFactory.fromHocon(origin, hoconConfig);
-    }
-
-    private static ConfigOrigin enrichOriginWithIncludes(FileConfigOrigin baseOrigin, com.typesafe.config.Config parsedConfig) {
-        var includedFiles = extractIncludedFiles(parsedConfig);
-        includedFiles.remove(baseOrigin.path().toAbsolutePath());
-        if (includedFiles.isEmpty()) {
-            return baseOrigin;
-        }
-        var origins = new ArrayList<ConfigOrigin>();
-        origins.add(baseOrigin);
-        for (var includedFile : includedFiles) {
-            if (Files.exists(includedFile)) {
-                origins.add(new FileConfigOrigin(includedFile));
-            }
-        }
-        if (origins.size() == 1) {
-            return baseOrigin;
-        }
-        return new ContainerConfigOrigin(origins);
-    }
-
-    /**
-     * Walks the parsed Typesafe Config tree and collects all unique file paths
-     * from which config values originated (including include files).
-     */
-    static Set<Path> extractIncludedFiles(com.typesafe.config.Config config) {
-        var files = new LinkedHashSet<Path>();
-        collectFileOrigins(config.root(), files);
-        return files;
-    }
-
-    private static void collectFileOrigins(com.typesafe.config.ConfigValue value, Set<Path> files) {
-        var origin = value.origin();
-        var filename = origin.filename();
-        // Typesafe Config 1.4.4 creates synthetic origins like "merge of file1, file2" when merging values
-        if (filename != null && !filename.startsWith("merge of ")) {
-            files.add(Path.of(filename).toAbsolutePath());
-        }
-        if (value instanceof com.typesafe.config.ConfigObject obj) {
-            for (var entry : obj.entrySet()) {
-                collectFileOrigins(entry.getValue(), files);
-            }
-        } else if (value instanceof com.typesafe.config.ConfigList list) {
-            for (var item : list) {
-                collectFileOrigins(item, files);
-            }
-        }
     }
 }
